move max-concurrency updates to v2
