+++ events.go
@@ -19,6 +19,4 @@ const (
 	CreatedBlobEvent       = "created-blob"
 	FailedEvent            = "failed"
 	ErrorEvent             = "error"
-	FlushStartEvent        = "flush-start"
-	FlushDoneEvent         = "flush-done"
 )
