+++ operation.go
@@ -1,8 +1,6 @@
 package batcher
 
-import (
-	"sync/atomic"
-)
+import "sync/atomic"
 
 type IOperation interface {
 	Payload() interface{}
