+++ v2/events.go
@@ -19,4 +19,6 @@ const (
 	CreatedBlobEvent       = "created-blob"
 	FailedEvent            = "failed"
 	ErrorEvent             = "error"
+	FlushStartEvent        = "flush-start"
+	FlushDoneEvent         = "flush-done"
 )
